Core and CLI: Rename modules
